Move custom JSON editor into a tab
